requirements.txt
requests==2.31.0
python-dotenv==1.0.0
anthropic==0.39.0
supabase==2.3.4
